Backend/requirements.txt
python==3.10
Flask==2.3.2
flask-cors==3.0.10
pymongo==4.6.1
pillow==10.3.0
pytz==2024.1
google-generativeai==0.5.1
requests
